classify_model/app/requirements.txt
redis
Pillow
torchvision
numpy
flask
torch
